refactor: 졸업 인증 기준 콘텐츠 별도 파일로 분리
